pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "easyppisp"
version = "0.1.0"
description = "A developer-friendly wrapper for Physically-Plausible ISP (PPISP) simulations"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "easyppisp contributors" }
]
keywords = ["image-processing", "ISP", "camera", "differentiable", "radiance-field", "nerf"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pillow>=10.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocstrings[python]>=0.24",
]

[project.scripts]
easyppisp = "easyppisp.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
easyppisp = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.pixi.project]
channels = ["conda-forge", "pytorch"]
platforms = ["linux-64"]

[tool.pixi.dependencies]
python = "3.10.*"
pytorch = ">=2.0.0"
torchvision = "*"
numpy = "*"
pytest = ">=7.4"
pillow = ">=10.0"

[tool.pixi.tasks]
build-cuda-docker = "docker run --rm -v /workspaces/easyppisp:/workspaces/easyppisp -w /workspaces/easyppisp nvidia/cuda:11.8.0-devel-ubuntu22.04 /bin/bash /workspaces/easyppisp/src/easyppisp/_internal/cuda/compile_cuda.sh"
